pyproject.toml
[project]
name = "s4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cmake>=4.1.2",
    "datasets>=4.3.0",
    "einops>=0.8.1",
    "gluonts>=0.16.2",
    "hydra-core>=1.3.2",
    "lit>=18.1.8",
    "marimo>=0.17.3",
    "matplotlib>=3.10.7",
    "numba>=0.62.1",
    "numpy>=2.1.3",
    "omegaconf>=2.3.0",
    "pandas>=2.3.3",
    "polars>=1.34.0",
    "pytorch-lightning==2.0.4",
    "rich>=14.2.0",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "sktime>=0.39.0",
    "timm==0.5.4",
    "torchtext>=0.18.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "wandb>=0.22.3",
]
